Wrap value in read-string
When using nrebl.middleare with Cursive, the values in the REBL where wrapped with double quotes. The value from nrepl whould be a readable string. This patch checks that the value is present (it might not be for all responese, in particular `:status #{:done}` has no `:value`.), then it will (read-string) on that before passing it to datafy.